requirements.txt
openai>=1.6.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.0.0
jupyter>=1.0.0
